requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
python-dotenv==0.19.0
Werkzeug==2.0.1
Authlib==0.15.5
requests==2.26.0
psycopg2-binary==2.9.3
SQLAlchemy==1.4.23
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
Supabase==0.0.2
OpenCV==4.5.3.56
Numpy==1.21.2
Pillow==8.3.2
OpenAI==1.3.0
clarifai==11.5.2
google-generativeai==0.3.2 